refactor: 프로젝트 구조 변경
- config / domain / feign / jwt 로 구조화
