+++ src/main/java/efub/eday/edayback/config/AuthenticationConfig.java
@@ -8,6 +8,8 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+import efub.eday.edayback.jwt.JwtExceptionFilter;
+import efub.eday.edayback.jwt.JwtFilter;
 import efub.eday.edayback.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
